+++ src/actions/types.js
@@ -4,3 +4,4 @@ export const FETCH_POSTS = 'FETCH_POSTS';
 export const NEW_POST = 'NEW_POST';
 export const SET_LOGIN_STATE = 'SET_LOGIN_STATE';
 export const LOGOUT = 'LOGOUT';
+export const STILL_LOGGEDIN = 'STILL_LOGGEDIN';
