+++ src/screens/stack/AddToShelfStack.js
@@ -1,0 +1,49 @@
+import React from 'react';
+import {StyleSheet, Text, View, TouchableOpacity} from 'react-native';
+import {createStackNavigator} from '@react-navigation/stack';
+import Ionicons from 'react-native-vector-icons/Ionicons';
+
+import AddShelf from '../dashboard/AddShelf';
+
+const Stack = createStackNavigator();
+
+const AddToShelfStack = ({navigation}) => {
+  return (
+    <Stack.Navigator
+      initialRouteName="AddShelf"
+      screenOptions={() => ({
+        headerTitleStyle: {
+          fontFamily: 'Nunito-Regular',
+          fontSize: 20,
+        },
+        headerBackTitleVisible: false,
+        headerTitleAlign: 'center',
+        headerTintColor: '#000',
+        headerTitleAlign: 'left',
+        title: 'Add To Shelf',
+        headerRight: () => {
+          return (
+            <TouchableOpacity
+              onPress={() => {
+                navigation.openDrawer();
+              }}>
+              <Ionicons
+                name="md-menu"
+                size={30}
+                style={{
+                  paddingRight: 20,
+                  fontFamily: 'Nunito-BoldItalic',
+                }}
+              />
+            </TouchableOpacity>
+          );
+        },
+      })}>
+      <Stack.Screen name="AddShelf" component={AddShelf} />
+    </Stack.Navigator>
+  );
+};
+
+export default AddToShelfStack;
+
+const styles = StyleSheet.create({});
